Fix payment CSS and site navigation
Properly loads the intended CSS for the makeaPayment page. It's not consistent with the other pages, which should probably be addressed, but the page is more readable now.

Also properly fixes the links in the header/dropdown menus, so that the pages can be comfortably navigated between (other than the "Inbox" page on the payment page, which doesn't exist)