Use Case 2 basic function
The back-end of paying off a loan is functional now.

When a loan application is submitted, that loan is added to the "outstanding loans" table. On the "make a payment" page, the loan applications table is checked to see if the user has applied for any loads. If they have, the outstanding loan table is checked for that loan, which is then loaded. The user can choose how much money to put towards the balance from their (imaginary) card. That amount is then subtracted from the balance.

TODO:
- proper landing page after submitting a payment
- proper handling for the user not having any outstanding loans
- allow the user to choose which of their loans to pay rather than finding the first available.